fix: downloader now supports cookies
Way to many youtube requests fail now due to the lack of cookies. This means I will be forced to implement a way to upload new cookies or get them automatically. For now though, when you save cookies at cookies.txt in the base dir of the app it will work